% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{f}
\alias{f}
\title{The probability density of standard Laplace distribution using Rcpp}
\usage{
f(x)
}
\arguments{
\item{x}{the value of the variable}
}
\value{
the probability density of the value of x
}
\description{
The probability density of standard Laplace distribution using Rcpp
}
\examples{
\dontrun{
f(4)
}
}
