% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunction.R
\name{rec_char}
\alias{rec_char}
\title{Reversal function}
\usage{
rec_char(x)
}
\arguments{
\item{x}{a vector of some positive integers}
}
\value{
the result of turning
}
\description{
A simple algorithm is used to turn the numbers.
}
\examples{
\dontrun{
char <- c('a','b','c','d')
rec_char(char)
}
}
