% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwMetropolis}
\alias{rwMetropolis}
\title{An Rcpp function for  generating the standard Laplace distribution}
\usage{
rwMetropolis(sigma, x0, N)
}
\arguments{
\item{sigma}{the param of Laplace distribution}

\item{x0}{the initial value}

\item{N}{the sample size}
}
\value{
a random sample of size \code{n}
}
\description{
An Rcpp function for  generating the standard Laplace distribution
}
\examples{
\dontrun{
rwMetropolis(.2,5,3000)
}
}
