\name{StatComp20076-package}
\alias{StatComp20076-package}
\alias{StatComp20076}
\docType{package}
\title{
  The final homework
}
\description{
  It is a simple R package developed to carry out two functions (implemented through the R package _Rcpp_) for the 'Statistical Computing' course.  
}
\details{
  Two functions are considered, namely, GetAcf and rec_char.
  The first vignettes file illustrates the usage of the (at least two) functions for my research project.
  The second one includes all my homework answers.
}
\author{
 Jinwen Wu<1292102295@qq.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
